examples/usecases/token_validation/pyproject.toml
[tool.poetry]
name = "token_validation"
version = "0.1.0"
description = "Validate JWTs against Keycloak"
authors = ["Coherent <team@coherent.global>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = {extras = ["crypto"], version = "^2.10.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
